==> scripts/data_prep/04_taxonomic_mapping.Rmd <==
---
title: "Map Birdlife and eBird taxonomy to Jetz"
output: html_notebook
bibliography: ../ref.bib 
---

```{r}
source('../env.R')
```

We map to Jetz as our working taxonomy to allow us to use a clade tree.

# Map Birdlife and eBird

## Compare missing species

### Read in Birdlife
```{r}
birdlife_tax_input = read_csv(filename(BIRDLIFE_WORKING_OUTPUT_DIR, 'taxonomy.csv'))
names(birdlife_tax_input) = c('birdlife_id', 'species_name', 'birdlife_common_name', 'alternative_name', 'status')
head(birdlife_tax_input)
```
```{r}
unique(birdlife_tax_input$status)
```

Drop extinct species
```{r}
birdlife_tax = birdlife_tax_input[birdlife_tax_input$status != 'EW',]
```

### Read in eBird
```{r}
ebird_tax = read_csv(filename(EBIRD_WORKING_OUTPUT_DIR, 'taxonomy.csv'))
names(ebird_tax) = c('ebird_id', 'ebird_category', 'ebird_species_fullname', 'ebird_common_name')
head(ebird_tax)
```

First create the root species name for all ebird records
```{r}
ebird_tax$species_name = word(ebird_tax$ebird_species_fullname, 1, 2, sep = ' ')
head(ebird_tax)
```

### Species in eBird but not Birdlife
```{r}
ebird_tax[!(ebird_tax$species_name %in% birdlife_tax$species_name),]
```

### Species in Birdlife but not eBird
```{r}
birdlife_tax[!(birdlife_tax$species_name %in% ebird_tax$species_name),]
```

### Expand birdlife alternative names
```{r}
birdlife_tax_alt_names = birdlife_tax[!is.na(birdlife_tax$alternative_name),c('birdlife_id', 'species_name', 'birdlife_common_name', 'alternative_name')] %>% separate_rows(alternative_name, sep = ";")
birdlife_tax_alt_names
```

Trying using Birdlife alternative name for finding matching eBird species
```{r}
birdlife_species_missing_from_ebird_with_alt_name = birdlife_tax_alt_names[!(birdlife_tax_alt_names$species_name %in% ebird_tax$species_name),]
inner_join(birdlife_species_missing_from_ebird_with_alt_name, ebird_tax, by = c('alternative_name' = 'species_name'))
```

## Build initial taxonomy mapping with straight forward species
```{r}
ebird_birdlife_taxonomy_mapping = inner_join(
  birdlife_tax[, c('species_name', 'birdlife_id', 'birdlife_common_name')], 
  ebird_tax[, c('species_name', 'ebird_id', 'ebird_common_name', 'ebird_species_fullname')]
)
ebird_birdlife_taxonomy_mapping$ebird_species_name = ebird_birdlife_taxonomy_mapping$species_name
head(ebird_birdlife_taxonomy_mapping)
```

```{r}
mapped_by_alternative_name = inner_join(
  birdlife_species_missing_from_ebird_with_alt_name[, c('species_name', 'birdlife_id', 'birdlife_common_name', 'alternative_name')], 
  ebird_tax[, c('species_name', 'ebird_id', 'ebird_common_name', 'ebird_species_fullname')],
  by = c('alternative_name' = 'species_name')
)
names(mapped_by_alternative_name) = c('species_name', 'birdlife_id', 'birdlife_common_name', 'ebird_species_name', 'ebird_id', 'ebird_common_name', 'ebird_species_fullname')
mapped_by_alternative_name = mapped_by_alternative_name[,c('species_name', 'birdlife_id', 'birdlife_common_name', 'ebird_id', 'ebird_common_name', 'ebird_species_fullname', 'ebird_species_name')]
mapped_by_alternative_name
```

```{r}
ebird_birdlife_taxonomy_mapping = rbind(
  ebird_birdlife_taxonomy_mapping,
  mapped_by_alternative_name
)
```


# Species left to map

## Species in Birdlife but in our new taxonomy
```{r}
missing_birdlife_species = birdlife_tax[!(birdlife_tax$species_name %in% ebird_birdlife_taxonomy_mapping$species_name),]
missing_birdlife_species
```


## Species in eBird but not our new taxonomy
```{r}
missing_ebird_species = ebird_tax[!(ebird_tax$species_name %in% ebird_birdlife_taxonomy_mapping$ebird_species_name),]
missing_ebird_species
```

### Map Remaining Birdlife Species
```{r}
ebird_tax$ebird_species_name = ebird_tax$species_name
```

### 60483076: Spilopelia suratensis(Western Spotted Dove)

* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=AEF9E7D341798053
* https://ebird.org/species/spodov
* http://datazone.birdlife.org/species/factsheet/western-spotted-dove-spilopelia-suratensis

```{r}
birdlife_spilopelia_suratensis = cbind(
  ebird_tax[ebird_tax$species_name == 'Spilopelia chinensis', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Spilopelia suratensis', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_spilopelia_suratensis)
birdlife_spilopelia_suratensis
```

### 22728712: Nesoenas duboisi(Reunion Pigeon)
Extinct, so exclude from taxonomy

### 22728289: Patagioenas albipennis (White-winged Pigeon)

* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=9DAE75390A64BCF5
* http://datazone.birdlife.org/species/factsheet/white-winged-pigeon-patagioenas-albipennis
* https://ebird.org/species/spwpig3/

```{r}
birdlife_patagioenas_albipennis = cbind(
  ebird_tax[ebird_tax$ebird_species_name == 'Patagioenas maculosa',c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Patagioenas albipennis', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_patagioenas_albipennis)
birdlife_patagioenas_albipennis
```

### 22725270: Patagioenas albilinea(Southern Band-tailed Pigeon)

* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=975A928B4D624F19
* http://datazone.birdlife.org/species/factsheet/southern-band-tailed-pigeon-patagioenas-albilinea
* https://ebird.org/species/batpig1/

```{r}
birdlife_patagioenas_albilinea = cbind(
  ebird_tax[ebird_tax$ebird_species_name == 'Patagioenas fasciata',c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Patagioenas albilinea', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_patagioenas_albilinea)
birdlife_patagioenas_albilinea
```

### 22690868: Leptotila battyi(Brown-backed Dove)

* Leptotila battyi: endemic to Panama, no records on eBird of this species or any Leptotila species in Panama.

### 60444823: Pampusana johannae(Eastern Bronze Ground-dove)

* https://avibase.bsc-eoc.org/species.jsp?avibaseid=A8329D59

```{r}
birdlife_otidiphaps_aruensis = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Pampusana beccarii [johannae Group]', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Pampusana johannae', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_otidiphaps_aruensis)
birdlife_otidiphaps_aruensis
```

### 22728031: Otidiphaps aruensis(White-naped Pheasant-pigeon)

* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=F197D25F1AB50C3B
* https://ebird.org/species/phepig1/

```{r}
birdlife_otidiphaps_aruensis = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Otidiphaps nobilis aruensis', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Otidiphaps aruensis', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_otidiphaps_aruensis)
birdlife_otidiphaps_aruensis
```

### 22725258: Otidiphaps cervicalis(Grey-naped Pheasant-pigeon)
```{r}
birdlife_otidiphaps_cervicalis = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Otidiphaps nobilis cervicalis', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Otidiphaps cervicalis', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_otidiphaps_cervicalis)
birdlife_otidiphaps_cervicalis
```

### 22726273: Otidiphaps insularis(Black-naped Pheasant-pigeon)
```{r}
birdlife_otidiphaps_insularis = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Otidiphaps nobilis insularis', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Otidiphaps insularis', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_otidiphaps_insularis)
birdlife_otidiphaps_insularis
```

### 22727834: Phapitreron nigrorum(Buff-eared Brown-dove)

```{r}
birdlife_phapitreron_nigrorum = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Phapitreron leucotis nigrorum', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Phapitreron nigrorum', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_phapitreron_nigrorum)
birdlife_phapitreron_nigrorum
```


### 22727841: Phapitreron brevirostris(Short-billed Brown-dove)

```{r}
birdlife_phapitreron_brevirostris = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Phapitreron leucotis brevirostris/occipitalis', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Phapitreron brevirostris', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_phapitreron_brevirostris)
birdlife_phapitreron_brevirostris
```


### 60543526: Phapitreron maculipectus(Grey-breasted Brown-dove)

```{r}
birdlife_phapitreron_maculipectus = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Phapitreron amethystinus maculipectus', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Phapitreron maculipectus', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_phapitreron_maculipectus)
birdlife_phapitreron_maculipectus
```


### 60543516: Phapitreron frontalis(Cebu Brown-dove)
```{r}
birdlife_phapitreron_frontalis = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Phapitreron amethystinus frontalis', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Phapitreron frontalis', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_phapitreron_frontalis)
birdlife_phapitreron_frontalis
```

### 45433495: Treron delalandii(Grey-breasted Green-pigeon)

* https://avibase.bsc-eoc.org/search.jsp?qstr=Treron+permagnus

```{r}
birdlife_treron_delalandii = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Treron calvus delalandii/granti', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Treron delalandii', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_treron_delalandii)
birdlife_treron_delalandii
```

### 22727551: Treron permagnus(Ryukyu Green-pigeon)    

* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=1EE907AC088D0B85

```{r}
birdlife_treron_permagnus = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Treron formosae permagnus/medioximus', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Treron permagnus', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_treron_permagnus)
birdlife_treron_permagnus
```

### 22725304: Ducula obiensis(Rusty Imperial-pigeon)   

* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=AAC237408FEB8125

```{r}
birdlife_ducula_obiensis = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Ducula basilica obiensis', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Ducula obiensis', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_ducula_obiensis)
birdlife_ducula_obiensis
```

### 22725071: Ramphiculus epius(Maroon-chinned Fruit-dove) 
```{r}
birdlife_ramphiculus_epius = cbind(
  ebird_tax[ebird_tax$ebird_species_name == 'Ptilinopus epius', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Ramphiculus epius', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_ramphiculus_epius)
birdlife_ramphiculus_epius
```


### 22725080: Ramphiculus mangoliensis(Sula Fruit-dove)    

```{r}
birdlife_ramphiculus_mangoliensis = cbind(
  ebird_tax[ebird_tax$ebird_species_name == 'Ptilinopus mangoliensis', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Ramphiculus mangoliensis', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_ramphiculus_mangoliensis)
birdlife_ramphiculus_mangoliensis
```


### 22691332: Ramphiculus fischeri(Red-eared Fruit-dove)
Ramphiculus meridionalis

```{r}
birdlife_ramphiculus_fischeri = cbind(
  ebird_tax[ebird_tax$ebird_species_name == 'Ptilinopus fischeri', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$birdlife_id == '22691332', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_ramphiculus_fischeri)
birdlife_ramphiculus_fischeri
```

### 22725286: Ptilinopus temminckii(Western Superb Fruit-dove)

* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=722E8494A2E5BD3C

```{r}
birdlife_ptilinopus_temminckii = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Ptilinopus superbus temminckii', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Ptilinopus temminckii', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_ptilinopus_temminckii)
birdlife_ptilinopus_temminckii
```


### 22728059: Ptilinopus gestroi(Eastern Ornate Fruit-dove)

* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=E0D24D45C31E80AB

```{r}
birdlife_ptilinopus_gestroi = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Ptilinopus ornatus gestroi', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Ptilinopus gestroi', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_ptilinopus_gestroi)
birdlife_ptilinopus_gestroi
```

### 22728271: Ptilinopus fasciatus(Samoan Fruit-dove)
Ptilinopus fasciatus

* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=9F3042D2364F20B7

```{r}
birdlife_ptilinopus_fasciatus = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Ptilinopus porphyraceus fasciatus', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Ptilinopus fasciatus', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_ptilinopus_fasciatus)
birdlife_ptilinopus_fasciatus
```
### 22690868	Leptotila battyi
* https://avibase.bsc-eoc.org/species.jsp?avibaseid=CC39ED207FDAB555

```{r}
ebird_tax[ebird_tax$ebird_species_fullname == 'Leptotila plumbeiceps', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')]
```

```{r}
birdlife_leptotila_battyi = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Leptotila plumbeiceps', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Leptotila battyi', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_leptotila_battyi)
birdlife_leptotila_battyi
```

### 22728146	Ramphiculus subgularis
Ptilinopus subgularis
* https://avibase.bsc-eoc.org/species.jsp?avibaseid=E5F5E7D2C7B69514

```{r}
birdlife_ramphiculus_subgularis = cbind(
  ebird_tax[ebird_tax$ebird_species_fullname == 'Ptilinopus subgularis', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Ramphiculus subgularis', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, birdlife_ramphiculus_subgularis)
birdlife_ramphiculus_subgularis
```

### Confirm all non-extinct species mapped
```{r}
birdlife_tax[!(birdlife_tax$species_name %in% ebird_birdlife_taxonomy_mapping$species_name),]
```

## eBird
## Species in eBird but not our new taxonomy
```{r}
missing_ebird_species = ebird_tax[!(ebird_tax$species_name %in% ebird_birdlife_taxonomy_mapping$ebird_species_name),]
missing_ebird_species
```

```{r}
ebird_species_extinct = c()
```

### 1925 Columba thiriouxi
Extinct
* https://avibase.bsc-eoc.org/species.jsp?avibaseid=B01378DFDE407C6D

```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Columba thiriouxi')
```

### 1938 Columba jouyi
Extinct
* https://avibase.bsc-eoc.org/species.jsp?avibaseid=33D2620F2FE88FC1

```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Columba jouyi')
```

### 1939 Columba versicolor
Extinct
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=0EFF2CA35EEEC5CB
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Columba versicolor')
```

### 2023 Ectopistes migratorius
Extinct passenger pigeon
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Ectopistes migratorius')
```

### 2033	Nesoenas cicur
Extinct
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=E54E8DE67BA6C994
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Nesoenas cicur')
```

### 2034 Nesoenas rodericanus
Extinct 
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=2BB91092BC7E5063
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Nesoenas rodericanus')
```

### 2119: Macropygia doreya(Sultan's Cuckoo-Dove)

2120: Macropygia doreya(Sultan's Cuckoo-Dove (Sulawesi))  

2125: Macropygia doreya(Sultan's Cuckoo-Dove (Sultan's))

* https://avibase.bsc-eoc.org/species.jsp?avibaseid=DF44EED5
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=FD5160925CD583C1

```{r}
ebird_macropygia_doreya = cbind(
  ebird_tax[ebird_tax$species_name == 'Macropygia doreya', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$birdlife_id == '22690561', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, ebird_macropygia_doreya)
ebird_macropygia_doreya
```

### 2143: Macropygia modiglianii(Barusan Cuckoo-Dove) 

* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=3A905A46&sec=summary&ssver=1

```{r}
ebird_macropygia_modiglianii = cbind(
  ebird_tax[ebird_tax$species_name == 'Macropygia modiglianii', c('ebird_id', 'ebird_species_name', 'ebird_species_fullname', 'ebird_common_name')],
  birdlife_tax[birdlife_tax$species_name == 'Macropygia emiliana', c('birdlife_id', 'species_name', 'birdlife_common_name')]
)
ebird_birdlife_taxonomy_mapping = rbind(ebird_birdlife_taxonomy_mapping, ebird_macropygia_modiglianii)
ebird_macropygia_modiglianii
```
### 2205 Pampusana salamonis	
Extinct 
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=71BE72FC64AB2A87
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Pampusana salamonis')
```

### 2206 Pampusana ferruginea
Extinct
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=8832822EC4F51ABA
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Pampusana ferruginea')
```

### 2224 Pampusana norfolkensis
Extinct
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=8D0BB075EC35B34B
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Pampusana norfolkensis')
```

### 2436	Zenaida graysoni
Extinct in the wild
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=5428D8973A682327
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Zenaida graysoni')
```

### 2440	Caloenas maculata
Extinct in the wild
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=5428D8973A682327
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Caloenas maculata')
```

### 2441	Raphus cucullatus
Extinct Dodo
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=DEF9592737DC3B5E
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Raphus cucullatus')
```

### 2442	Pezophaps solitaria
Extinct
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=2FC54A500BFB61BA
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Pezophaps solitaria')
```

### 2464	Microgoura meeki
Extinct
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=75D098B55481BD7E
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Microgoura meeki')
```

### 2674	Ptilinopus mercierii
Extinct
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=B526D1AE92AC6811
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Ptilinopus mercierii')
```

### 2738	Alectroenas nitidissimus
Extinct
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=BE8566B00A06FE64
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Alectroenas nitidissimus')
```

### 2739	Alectroenas payandeei
Extinct
* https://avibase.bsc-eoc.org/species.jsp?lang=EN&avibaseid=05E1097F65FC6090
```{r}
ebird_species_extinct = append(ebird_species_extinct, 'Alectroenas payandeei')
```

### Confirm all non-extinct species mapped
```{r}
ebird_tax[!(ebird_tax$ebird_species_name %in% ebird_birdlife_taxonomy_mapping$ebird_species_name) & !(ebird_tax$ebird_species_name %in% ebird_species_extinct),]
```


# Result
```{r}
ebird_birdlife_taxonomy_mapping
```


## Store Mapping file
```{r}
write_csv(ebird_birdlife_taxonomy_mapping, filename(TAXONOMY_OUTPUT_DIR, 'taxonomy_mapping.csv'))
```

## Create phylogeny
```{r}
pruned_tree = extractTree(species = ebird_birdlife_taxonomy_mapping$ebird_species_name, output.type = 'scientific', taxonomy.year = '2023', version = 'current')
write.tree(pruned_tree, filename(TAXONOMY_OUTPUT_DIR, 'phylogeny.tre'))
plot(pruned_tree)
```

```{r}
getCitations(pruned_tree)
```


# Columbidae family overvide
```{r}
distinct(ebird_birdlife_taxonomy_mapping[,c('species_name')])
```
Total genera
```{r}
ebird_birdlife_taxonomy_mapping %>% 
  mutate(family = gsub( " .*$", "", species_name)) %>%
  dplyr::select(family) %>%
  distinct()
```
