==> scripts/data_prep/05_generate_abundance_test_data.Rmd <==
---
title: "Generate presence test data using eBird Status & Trends"
output: html_notebook
bibliography: ../ref.bib 
---

```{r}
source('../env.R')
```

# Select species
Load the ebird taxonomy

```{r}
ebird_tax = read_csv(filename(EBIRD_WORKING_OUTPUT_DIR, 'taxonomy.csv'))
names(ebird_tax) = c('ebird_id', 'category', 'species_name', 'common_name')
head(ebird_tax)
```

Filter against library list, as not everything on website is availble in library
```{r}
species_available = ebirdst_runs$scientific_name[ebirdst_runs$scientific_name %in% ebird_tax$species_name]
paste(species_available[order(species_available)], collapse=", ")
```

# Download Status and Trends data
```{r}
key = read_file(filename(KEYS_DIR, 'ebirdst.txt'))
set_ebirdst_access_key(key, overwrite = T)
ebirdst_data_dir()
```

```{r}
for (species in species_available) {
  print(paste('fetching', species))
  ebirdst_download_status(
    species = species,
    download_abundance = TRUE,
    download_occurrence = FALSE,
    download_count = FALSE,
    download_ranges = FALSE,
    download_regional = FALSE,
    download_pis = FALSE,
    download_ppms = FALSE,
    download_all = FALSE
  )
}
```

## Create abundance data
Read in our initial city vectors, we will join these to our checklists
```{r}
initial_city_selection = vect(filename(mkdir(GEO_WORKING_OUTPUT_DIR, 'cities'), 'initial_selection.shp'))
```

```{r}
city_data = data.frame(initial_city_selection)
city_data$ID = as.double(rownames(city_data))
city_data
```

```{r}
abundance_data = data.frame()

for (species in species_available) {
  abd <- load_raster(species, resolution = "3km", product = "abundance", period = "seasonal")
  abd_wgs84 = project(abd, initial_city_selection)
  city_abundance = terra::extract(abd_wgs84, initial_city_selection)
  
  if ("resident" %in% colnames(city_abundance)) {
     result = city_abundance %>% group_by(ID) %>% summarise(
       mean_breeding = NA, 
       median_breeding = NA, 
       sd_breeding = NA,
       mean_nonbreeding = NA, 
       median_nonbreeding = NA, 
       sd_nonbreeding = NA,
       mean_resident = mean(resident), 
       median_resident = median(resident), 
       sd_resident = sd(resident), 
       species = species
     )
     abundance_data = rbind(abundance_data, inner_join(result, city_data[,c('ID', 'city_id', 'city_name')]))
  } else {
    result = city_abundance %>% group_by(ID) %>% summarise(
      mean_breeding = mean(breeding), 
      median_breeding = median(breeding), 
      sd_breeding = sd(breeding), 
      mean_nonbreeding = mean(nonbreeding), 
      median_nonbreeding = median(nonbreeding), 
      sd_nonbreeding = sd(nonbreeding), 
      mean_resident = NA, 
      median_resident = NA, 
      sd_resident = NA,
      species = species
    ) 
    abundance_data = rbind(abundance_data, inner_join(result, city_data[,c('ID', 'city_id', 'city_name')]))
  }
}

head(abundance_data)
```

```{r}
write_csv(abundance_data, filename(EBIRD_WORKING_OUTPUT_DIR, 'ebird_trends_abundance_test_data.csv'))
abundance_data = read_csv(filename(EBIRD_WORKING_OUTPUT_DIR, 'ebird_trends_abundance_test_data.csv'))
```
