==> scripts/data_prep/06_create_urban_pools.Rmd <==
---
title: "Find parameters for building urban communities"
output: html_notebook
bibliography: ../ref.bib 
---

Using eBird trends and status abundance data, can we find a minimum number of city total checklists required, and a cut off percentage of checklists recorded, such that we can be confident that any species recorded on more than that percentage of checklist is present in the city (given the city has more that the total minimum total checklists recorded within it).

```{r}
source('../env.R')
options(warn=-1)
```

```{r}
taxonomic_mapping = read_csv(filename(TAXONOMY_OUTPUT_DIR, 'taxonomy_mapping.csv'))
```

```{r}
mapping_birdlife_to_ebird= unique(taxonomic_mapping[,c('species_name', 'ebird_species_name')])
mapping_birdlife_to_ebird
```



# Create regional pool data
This data is created from Birdlife (and some eBird) distrbution data, a species is recorded as occuring in a regional pool when its range polygon intersects with a city polygon.

A set of city names we will use for inspecting data:
```{r}
inspection_cities = c('Manchester', 'Bogota', 'Los Angeles', 'Jakarta', 'Nairobi', 'Medellín', 'San Jose')
```

Load data
```{r}
regional_pool_data_input = read_csv(filename(BIRDLIFE_WORKING_OUTPUT_DIR, 'urban_distributions.csv'))
regional_pool_data_input$seasonal = factor(regional_pool_data_input$seasonal, levels = c('Resident', 'Breeding Season', 'Non-breeding Season', 'Passage', ordered = T))
regional_pool_data_input$origin = factor(regional_pool_data_input$origin, levels = c('Native', 'Introduced', ordered = T))
regional_pool_data_input$presence = factor(regional_pool_data_input$presence, levels = c('Extant', 'Possibly Extant', 'Possibly Extinct', ordered = T))
regional_pool_data_input[regional_pool_data_input$city_name %in% inspection_cities,]
```

Map taxonomy to eBird
```{r}
regional_pool_data_seasonal = regional_pool_data_input %>% 
  left_join(mapping_birdlife_to_ebird) %>%
  arrange(seasonal) %>%
  group_by(city_id, city_name, ebird_species_name) %>%
  summarise(seasonal = first(seasonal)) %>%
  arrange(city_id, ebird_species_name)

regional_pool_data_origin = regional_pool_data_input %>% 
  left_join(mapping_birdlife_to_ebird) %>%
  arrange(origin) %>%
  group_by(city_id, city_name, ebird_species_name) %>%
  summarise(origin = first(origin)) %>%
  arrange(city_id, ebird_species_name)

regional_pool_data_presence = regional_pool_data_input %>% 
  left_join(mapping_birdlife_to_ebird) %>%
  arrange(presence) %>%
  group_by(city_id, city_name, ebird_species_name) %>%
  summarise(presence = first(presence)) %>%
  arrange(city_id, ebird_species_name)

regional_pool_data = inner_join(regional_pool_data_seasonal, inner_join(regional_pool_data_origin, regional_pool_data_presence))
```

Final regional pool data
```{r}
regional_pool_data[regional_pool_data$city_name %in% inspection_cities,]
```

# Create urban summary data
Get number of checklists and total effort in each city
```{r}
urban_checklists = read_csv(filename(EBIRD_WORKING_OUTPUT_DIR, 'urban_checklists.csv'))

urban_summary = urban_checklists %>% group_by(city_id, city_name) %>% summarise(total_city_checklists = n(), total_city_effort = sum(duration), total_city_locations = n_distinct(locality_id))

urban_summary[urban_summary$city_name %in% inspection_cities,]
```

# Create urban pools
For each of our test species, find the number of checklists and total effort in each city from the eBird raw observational data
```{r}
ebird_raw = read_csv(filename(EBIRD_WORKING_OUTPUT_DIR, 'urban_columbidae.csv'))
head(ebird_raw)
```

Now calculate the total number of checklists each species appears on, and then total effort of those checklists within each city.
This gives a single row for each species in each city.
Then join the data onto the regional pool data, and the urban summary data.
```{r}
ebird_species_name_to_full_species_name = distinct(taxonomic_mapping[,c('ebird_species_fullname', 'ebird_species_name')])

urban_columbidae_presence_records = ebird_raw %>% 
  rename(ebird_species_fullname = species_name) %>%
  left_join(urban_checklists[,c('checklist_id', 'locality_id')]) %>%
  left_join(ebird_species_name_to_full_species_name) %>%
  group_by(city_id, city_name, ebird_species_name) %>% 
  summarise(total_presence_checklists = n(), total_presence_effort = sum(duration), total_presence_locations = n_distinct(locality_id)) %>%
  left_join(regional_pool_data, by = c('city_id' = 'city_id', 'city_name' = 'city_name', 'ebird_species_name' = 'ebird_species_name')) %>%
  left_join(urban_summary, by = c('city_id' = 'city_id', 'city_name' = 'city_name')) %>%
  arrange(city_id, ebird_species_name)
```

NA for seasonal/origin/presence indicates that the species is on an eBird checklist for the city, but not in the birdlife regional pool.
```{r}
urban_columbidae_presence_records[urban_columbidae_presence_records$city_name %in% inspection_cities,]
```

We also want records for each city that could have that pigeon (e.g. in regional pool range) but has no eBird records.
Create absence records for all species in the birdlife regional pool data for all cities.
```{r}
urban_columbidae_absence_records_part1 = left_join(regional_pool_data, urban_summary)

urban_columbidae_absence_records_part1$total_presence_checklists = 0
urban_columbidae_absence_records_part1$total_presence_effort = 0
urban_columbidae_absence_records_part1$total_presence_locations = 0

urban_columbidae_absence_records = urban_columbidae_absence_records_part1[!is.na(urban_columbidae_absence_records_part1$total_city_checklists),]

urban_columbidae_absence_records[urban_columbidae_absence_records$city_name %in% inspection_cities,]
```

Append the absence and presence data together, and then deduplicate to ensure only one species per city.
Then calculate the percentage of all checklists, and the percentage of total effort in a city for each species.
```{r}
urban_columbidae_presence_and_absence = rbind(urban_columbidae_presence_records, urban_columbidae_absence_records) %>%
  group_by(city_id, city_name, ebird_species_name, total_city_checklists, total_city_locations, total_city_effort, seasonal, presence, origin) %>%
  summarise(total_presence_checklists = max(total_presence_checklists), total_presence_effort = max(total_presence_effort), total_presence_locations = max(total_presence_locations)) %>%
  arrange(city_id, ebird_species_name)

urban_columbidae_presence_and_absence$percentage_checklists = urban_columbidae_presence_and_absence$total_presence_checklists * 100 / urban_columbidae_presence_and_absence$total_city_checklists
urban_columbidae_presence_and_absence$percentage_effort = urban_columbidae_presence_and_absence$total_presence_effort * 100 / urban_columbidae_presence_and_absence$total_city_effort
urban_columbidae_presence_and_absence$percentage_locations = urban_columbidae_presence_and_absence$total_presence_locations * 100 / urban_columbidae_presence_and_absence$total_city_locations

urban_columbidae_presence_and_absence[urban_columbidae_presence_and_absence$city_name %in% inspection_cities,]
```

```{r}
write_csv(urban_columbidae_presence_and_absence, filename(COMMUNITY_OUTPUT_DIR, 'recorded_species.csv'))
```

# Which data get removed when we only keep species in regional pool
If the birdlife vector does not intersect with the city, then we should remove the species from the urban pool - it does not make sense having a species in the urban but not the regional pool.
What records are we losing?
```{r}
data.frame(urban_columbidae_presence_and_absence) %>%
  filter(is.na(seasonal)) %>%
  dplyr::select(city_id, city_name, ebird_species_name, percentage_checklists, total_city_checklists) %>%
  arrange(percentage_checklists)
```


```{r}
data.frame(urban_columbidae_presence_and_absence) %>%
  filter(is.na(seasonal)) %>%
  filter(percentage_checklists > 1) %>%
  group_by(ebird_species_name) %>%
  summarise(total_cities_missing_in_regional_pool = n(), avg_percentage_checklists = mean(percentage_checklists)) %>%
  arrange(total_cities_missing_in_regional_pool)
```
