==> scripts/data_prep/08_generate_traits.Rmd <==
---
title: "Trait generation"
output: html_notebook
bibliography: ../ref.bib 
---

```{r}
source('../env.R')
```

# Create measured traits database for Jetz Columbidae Taxonomy
Here we create traits based on the @AVONET2022 morphology and trait database.
Avonet provide traits using the eBird taxonomy; however, it is based on the 2021 version while we're using the 2023 version.


## Load Avonet Traits
```{r}
avonet_ebird = read_csv(DL_AVONET)
avonet_ebird
```

### Avonet Columbidae traits
```{r}
avonet_columbidae = avonet_ebird[avonet_ebird$Order2 == 'Columbiformes',]
avonet_columbidae
```

### Check we have all species from our taxonomy file

```{r}
ebird_taxonomy_2021 = read_csv(DL_EBIRD_TAXONOMY_2021)
ebird_taxonomy_2023 = read_csv(DL_EBIRD_TAXONOMY)
```

### Check which species we are losing from avonet
Avonet uses eBird 2021 taxonomy, check all species are present.
```{r}
avonet_columbidae[!(avonet_columbidae$Species2 %in% ebird_taxonomy_2021$SCI_NAME),]
```

But these species are not in the 2023 taxonomy
```{r}
missing_species_names_from_2023_taxonomy = avonet_columbidae[!(avonet_columbidae$Species2 %in% ebird_taxonomy_2023$SCI_NAME),c('Species2')]
missing_species_names_from_2023_taxonomy
```

Fetch species code from 2021 taxonomy
```{r}
missing_species_codes_from_2023_taxonomy = ebird_taxonomy_2021[ebird_taxonomy_2021$SCI_NAME %in% missing_species_names_from_2023_taxonomy$Species2,c('SPECIES_CODE', 'SCI_NAME')]
names(missing_species_codes_from_2023_taxonomy) = c('SPECIES_CODE', 'SCI_NAME_2021')
missing_species_codes_from_2023_taxonomy
```

Which now map to:
```{r}
missing_taxa_mapping = left_join(missing_species_codes_from_2023_taxonomy, ebird_taxonomy_2023[ebird_taxonomy_2023$SPECIES_CODE %in% missing_species_codes_from_2023_taxonomy$SPECIES_CODE,c('SCI_NAME', 'SPECIES_CODE')])
names(missing_taxa_mapping) = c('SPECIES_CODE', 'SCI_NAME_2021', 'SCI_NAME_2023')
missing_taxa_mapping
```
Create dummy mapping for remaining species that are present in 2021 and 2023 taxonomy
```{r}
taxa_mapping = ebird_taxonomy_2023[ebird_taxonomy_2023$ORDER == 'Columbiformes' & !(ebird_taxonomy_2023$SCI_NAME %in% missing_taxa_mapping$SCI_NAME_2023), c('SPECIES_CODE', 'SCI_NAME')]
names(taxa_mapping) = c('SPECIES_CODE', 'SCI_NAME_2023')
taxa_mapping$SCI_NAME_2021 = taxa_mapping$SCI_NAME_2023

taxa_mapping = rbind(taxa_mapping, missing_taxa_mapping)
taxa_mapping
```

## Fetch required traits
```{r}
avonet_columbidae_ebird_2023 = avonet_columbidae %>%
  left_join(taxa_mapping, by = c('Species2' = 'SCI_NAME_2021')) %>%
  group_by(SCI_NAME_2023) %>%
  dplyr::select(
    SCI_NAME_2023, Beak.Width, `Hand-Wing.Index`, Mass, Habitat, Habitat.Density, Migration, Trophic.Level, Trophic.Niche, Primary.Lifestyle) %>%
  dplyr::rename(
    ebird_species_name = SCI_NAME_2023,
    beak_width = Beak.Width,
    hwi = `Hand-Wing.Index`,
    mass = Mass,
    habitat = Habitat,
    habitat_density = Habitat.Density,
    migration = Migration,
    trophic_level = Trophic.Level,
    trophic_niche = Trophic.Niche,
    primary_lifestyle = Primary.Lifestyle
  )

avonet_columbidae_ebird_2023
```

# Store our new trait data
Here we select the second axis of each trait PCA.
```{r}
write_csv(avonet_columbidae_ebird_2023, filename(TAXONOMY_OUTPUT_DIR, 'traits_ebird.csv'))
```

# Check which single traits are associated with mass
```{r}
ggplot(avonet_columbidae_ebird_2023, aes(x = mass, y = beak_width)) + geom_point()
```

```{r}
summary(lm(beak_width ~ mass, data = avonet_columbidae_ebird_2023))
```

```{r}
ggplot(avonet_columbidae_ebird_2023, aes(x = mass, y = hwi)) + geom_point()
```

```{r}
summary(lm(hwi ~ mass, data = avonet_columbidae_ebird_2023))
```


# Confirm we have a trait for all our urban species
```{r}
ebird_columbidae = read_csv(filename(EBIRD_WORKING_OUTPUT_DIR, 'urban_columbidae.csv'))
unique_species_names = ebird_columbidae %>% 
  group_by(species_name) %>% 
  dplyr::select(species_name) %>% 
  distinct %>% 
  ungroup
nrow(unique_species_names)
```
```{r}
mapping = read_csv(filename(TAXONOMY_OUTPUT_DIR, 'taxonomy_mapping.csv')) %>% dplyr::select('ebird_species_name', 'ebird_species_fullname') %>% distinct
species_mapped_to_simple_name = left_join(unique_species_names, mapping, by = c('species_name' = 'ebird_species_fullname'))
nrow(species_mapped_to_simple_name)
```
Missing single species
```{r}
species_mapped_to_simple_name[!(species_mapped_to_simple_name$ebird_species_name %in% avonet_columbidae_ebird_2023$ebird_species_name),]
```
Highly related to Ducula badia
* https://avibase.bsc-eoc.org/species.jsp?avibaseid=A938A06D6CE3CB1D
So create duplicate row

```{r}
ducula_cuprea = avonet_columbidae_ebird_2023[avonet_columbidae_ebird_2023$ebird_species_name == 'Ducula badia',]
ducula_cuprea$ebird_species_name = 'Ducula cuprea'
ducula_cuprea
```
# Store our updated trait data
Here we select the second axis of each trait PCA.
```{r}
write_csv(rbind(avonet_columbidae_ebird_2023, ducula_cuprea), filename(TAXONOMY_OUTPUT_DIR, 'traits_ebird.csv'))
```
