==> scripts/data_prep/10_select_thresholds.Rmd <==
---
title: "Select thresholds for urban communities"
output: html_notebook
bibliography: ../ref.bib  
---

```{r}
source('../env.R')
```

# Species investigation
@Bonier2007 contains a list of species considered to be urban birds. The species from the Columbiae family are:

* Columba fasciata (cities 70)
* Columba livia (cities 1,26,27,28,35,36,39,40, 47,54,62,66,67,72)
* Columba palumbus (cities 30,31,33,47,62)
* Columba picazuro (cities 6,11)
* Columbina cruziana (cities 56)
* Columbina inca (cities 29,69)
* Columbina passerina (cities 9)
* Columbina squammata (cities 6,8)
* Columbina talpacoti (cities 6,8,10,11,21,29,73)
* Streptopelia chinensis  (cities 17,18,37,42,61)    
* Streptopelia decaocto (cities 35,47,48,54,55,62,65,66,67)
* Streptopelia lugens (cities 32)
* Streptopelia orientalis (cities 14,18,51,52,55)
* Streptopelia semitorquata (cities 63)
* Streptopelia senegalensis (cities 26,27,28,47,53,54,60,66,67,72)

The species they consider to be rural birds are:

* Columba speciosa
* Columba oenas
* Columba cayennensis
* Columba plumbea
* Columbina cyanopis

Here will will investigate what each of these species look like within our pool data.

```{r}
pools = read_csv(filename(COMMUNITY_OUTPUT_DIR, 'recorded_species_with_spatial_data.csv'))
pools
```
```{r}
mapping = read_csv(filename(TAXONOMY_OUTPUT_DIR, 'taxonomy_mapping.csv'))
```
### City List from @Bonier2007
```{r}
bornier_cities = read_tsv(filename(PROVIDED_DATA, 'bornier_2007_cities.tsv'))
bornier_cities
```
```{r}
bornier_cities[!(bornier_cities$City %in% pools$city_name),]
```

```{r}
reported_rows = c('city_id', 'city_name', 'ebird_species_name', 'percentage_checklists', 'percentage_locations', 'total_city_checklists', 'total_city_locations', 'buf_500m_percentage_survey_area_present', 'buf_500m_percentage_total_city_area_surveyed')
```

## Columba fasciata 
```{r}
cities = bornier_cities$City[bornier_cities$`City Number` == 70]
Columba_fasciata = pools %>% 
  filter(ebird_species_name %in% c('Patagioenas fasciata')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Columba_fasciata                                                                                        
```

## Columba livia
Cities: 
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(1,26,27,28,35,36,39,40,47,54,62,66,67,72)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(1,26,27,28,35,36,39,40,47,54,62,66,67,72)]
Columba_livia = pools %>% 
  filter(ebird_species_name %in% c('Columba livia')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Columba_livia
```


## Columba palumbus
Cities: 
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(30,31,33,47,62)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(30,31,33,47,62)]
Columba_palumbus = pools %>% 
  filter(ebird_species_name %in% c('Columba palumbus')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Columba_palumbus
```

## Columba picazuro
Cities: 
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(6,11)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(6,11)]
Columba_picazuro = pools %>% 
  filter(ebird_species_name %in% c('Patagioenas picazuro')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Columba_picazuro
```
## Columbina cruziana
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(56)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(56)]
Columbina_cruziana = pools %>% 
  filter(ebird_species_name %in% c('Columbina cruziana')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Columbina_cruziana
```

## Columbina inca
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(29,69)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(29,69)]
Columbina_inca = pools %>% 
  filter(ebird_species_name %in% c('Columbina inca')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Columbina_inca
```

## Columbina passerina
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(9)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(9)]
Columbina_passerina = pools %>% 
  filter(ebird_species_name %in% c('Columbina passerina')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Columbina_passerina
```

## Columbina squammata (cities 6,8)
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(6,8)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(6,8)]
Columbina_squammata = pools %>% 
  filter(ebird_species_name %in% c('Columbina squammata')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Columbina_squammata
```

## Columbina talpacoti
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(6,8,10,11,21,29,73)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(6,8,10,11,21,29,73)]
Columbina_talpacoti = pools %>% 
  filter(ebird_species_name %in% c('Columbina talpacoti')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Columbina_talpacoti
```

## Streptopelia chinensis
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(17,18,37,42,61)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(17,18,37,42,61)]
Streptopelia_chinensis = pools %>% 
  filter(ebird_species_name %in% c('Spilopelia chinensis')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Streptopelia_chinensis
```
## Streptopelia decaocto
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(35,47,48,54,55,62,65,66,67)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(35,47,48,54,55,62,65,66,67)]
Streptopelia_decaocto = pools %>% 
  filter(ebird_species_name %in% c('Streptopelia decaocto')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Streptopelia_decaocto
```

## Streptopelia lugens
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(32)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(32)]
Streptopelia_lugens = pools %>% 
  filter(ebird_species_name %in% c('Streptopelia lugens')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Streptopelia_lugens
```

## Streptopelia orientalis
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(14,18,51,52,55)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(14,18,51,52,55)]
Streptopelia_orientalis = pools %>% 
  filter(ebird_species_name %in% c('Streptopelia orientalis')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Streptopelia_orientalis
```

## Streptopelia semitorquata
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(63)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(63)]
Streptopelia_semitorquata = pools %>% 
  filter(ebird_species_name %in% c('Streptopelia semitorquata')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Streptopelia_semitorquata
```

## Streptopelia senegalensis
```{r}
bornier_cities$City[bornier_cities$`City Number` %in% c(26,27,28,47,53,54,60,66,67,72)]
```

```{r}
cities = bornier_cities$City[bornier_cities$`City Number` %in% c(26,27,28,47,53,54,60,66,67,72)]
Streptopelia_senegalensis = pools %>% 
  filter(ebird_species_name %in% c('Spilopelia senegalensis')) %>% 
  filter(city_name %in% cities) %>% 
  dplyr::select(reported_rows)
Streptopelia_senegalensis
```
## Data Exploration
```{r}
all_species = bind_rows(Columba_fasciata,Columba_livia,Columba_palumbus,Columba_picazuro,Columbina_cruziana,Columbina_inca,Columbina_passerina,Columbina_squammata,Columbina_talpacoti,Streptopelia_chinensis,Streptopelia_decaocto,Streptopelia_lugens,Streptopelia_orientalis,Streptopelia_semitorquata,Streptopelia_senegalensis)
all_species %>% arrange(percentage_checklists)
```
Checklists
```{r}
all_species %>% 
  filter(total_city_checklists > 200) %>%
  summarise(
    number_records = n(),
    mean_percentage_checklists = mean(percentage_checklists),
    lower_10_quant_percentage_checklists = quantile(percentage_checklists, 0.1),
    lower_20_quant_percentage_checklists = quantile(percentage_checklists, 0.2),
    min_percentage_checklists = min(percentage_checklists),
    mean_percentage_locations = mean(percentage_locations),
    lower_10_quant_percentage_locations = quantile(percentage_locations, 0.1),
    lower_20_quant_percentage_locations = quantile(percentage_locations, 0.2),
    min_percentage_locations = min(percentage_locations)
  )
```

Locations
```{r}
all_species %>% 
  filter(total_city_locations > 100) %>%
  summarise(
    number_records = n(),
    mean_percentage_locations = mean(percentage_locations),
    lower_10_quant_percentage_locations = quantile(percentage_locations, 0.1),
    lower_20_quant_percentage_locations = quantile(percentage_locations, 0.2),
    min_percentage_locations = min(percentage_locations)
  )
```

Area
```{r}
all_species %>% 
  filter(buf_500m_percentage_total_city_area_surveyed > 10) %>%
  summarise(
    number_records = n(),
    mean_percentage_surveyed_area = mean(buf_500m_percentage_survey_area_present),
    lower_10_quant_percentage_surveyed_area = quantile(buf_500m_percentage_survey_area_present, 0.1),
    lower_20_quant_percentage_surveyed_area = quantile(buf_500m_percentage_survey_area_present, 0.2),
    min_percentage_surveyed_area = min(buf_500m_percentage_survey_area_present),
    mean_percentage_checklists = mean(percentage_checklists),
    lower_10_quant_percentage_checklists = quantile(percentage_checklists, 0.1),
    lower_20_quant_percentage_checklists = quantile(percentage_checklists, 0.2),
    min_percentage_checklists = min(percentage_checklists)
  )
```
