==> scripts/data_prep/11_species_city_distance_to_distribution_limits.Rmd <==
---
title: "Distribution limits"
output: html_notebook
---

```{r}
source('../env.R')
```

Here we wish to find out how close each city is to a species distribution limits.

# Load data
```{r}
initial_city_selection = st_read(filename(mkdir(GEO_WORKING_OUTPUT_DIR, 'cities'), 'initial_selection.shp'))
st_crs(initial_city_selection)
```

```{r}
birdlife = st_read(DL_BIRDLIFE_DISTRIBUTIONS)
st_crs(birdlife)
```


```{r}
pools = read_csv(filename(COMMUNITY_OUTPUT_DIR, 'jetz_all_recorded_species.csv'))
pools
```

## Calculate distance
```{r}
distribution_bounding_box = function(birdlife_species_names) {
  distributions = birdlife[birdlife$SCI_NAME %in% birdlife_species_names,]
  st_bbox(distributions)
}

distribution_northern_edge = function(birdlife_species_names, jetz_species_name) {
  distribution_bbox = distribution_bounding_box(birdlife_species_names)
  distribution_northern_limit = st_bbox(c(xmin = as.double(distribution_bbox['xmin']), xmax = as.double(as.double(distribution_bbox['xmax'])), ymax = as.double(distribution_bbox['ymax']) + 0.1, ymin = as.double(distribution_bbox['ymax'])), crs = st_crs(distribution_bbox))
  result = st_sf(geometry = st_as_sfc(distribution_northern_limit))
  result$SCI_NAME = jetz_species_name
  result
}

distribution_southern_edge = function(birdlife_species_names, jetz_species_name) {
  distribution_bbox = distribution_bounding_box(birdlife_species_names)
  distribution_southern_limit = st_bbox(c(xmin = as.double(as.double(distribution_bbox['xmin'])), xmax = as.double(as.double(distribution_bbox['xmax'])), ymax = as.double(distribution_bbox['ymin']), ymin = as.double(distribution_bbox['ymin']) - 0.1), crs = st_crs(distribution_bbox))
 result = st_sf(geometry = st_as_sfc(distribution_southern_limit))
 result$SCI_NAME = jetz_species_name
 result
}
```

```{r}
plot_distribution_limits = function(city_id, species_name) {
  city = initial_city_selection[initial_city_selection$city_id == city_id,]

  distribution_southern_limit_sf = distribution_southern_edge(c(species_name), species_name)
  distribution_northern_limit_sf = distribution_northern_edge(c(species_name), species_name)
  
  ggplot() +  
    geom_sf(data = distribution_southern_limit_sf, aes(geometry = geometry), colour = "blue") +
    geom_sf(data = distribution_northern_limit_sf, aes(geometry = geometry), colour = "green") +
    geom_sf(data = city, aes(geometry = geometry), colour = "red")
}
```

```{r}
plot_distribution_limits(10, 'Columbina passerina')
```

```{r}
plot_distribution_limits(624, 'Geotrygon montana')
```

```{r}
taxonomic_mapping = read_csv(filename(TAXONOMY_OUTPUT_DIR, 'taxonomy_mapping.csv'))
birdlife_to_jetz = taxonomic_mapping[, c('species_name', 'jetz_species_name')]
birdlife_to_jetz
```

```{r, warning=F}
unique_jetz_species_names = unique(birdlife_to_jetz$jetz_species_name)

birdlife_species_names = function(jetz_species_name) {
  birdlife_to_jetz$species_name[birdlife_to_jetz$jetz_species_name == jetz_species_name]
}

distribution_southern_edges = foreach(js_name = unique_jetz_species_names, .combine = 'rbind') %do% {
  species_names = birdlife_species_names(js_name)
  distribution_southern_edge(species_names, js_name)
}

distribution_northern_edges = foreach(js_name = unique_jetz_species_names, .combine = 'rbind') %do% {
  species_names = birdlife_species_names(js_name)
  distribution_northern_edge(species_names, js_name)
}
```

```{r}
city_in_distribution = function(city_id, jetz_species_name) {
  city = initial_city_selection[initial_city_selection$city_id %in% city_id,]
  species_names = birdlife_species_names(jetz_species_name)
  distribution_bbox = distribution_bounding_box(species_names)
  distribution = st_sf(geometry = st_as_sfc(distribution_bbox))
  lengths(st_intersects(distribution, city)) > 0
}

city_in_distribution(624, 'Geotrygon_montana')
city_in_distribution(1786, 'Geotrygon_montana')
```

```{r}
distance_to_edge = function(city_id, edge) {
  city = initial_city_selection[initial_city_selection$city_id %in% city_id,]
  city_centre = st_centroid(city)
  city_centre_bbox = st_bbox(edge)
  
  edge_bbox = st_bbox(edge)
  edge_point = st_bbox(c(xmin = as.double(city_centre_bbox['xmin']), xmax = as.double(as.double(city_centre_bbox['xmax'])), ymax = as.double(edge_bbox['ymax']), ymin = as.double(edge_bbox['ymax'])), crs = st_crs(city_centre))
  
  edge_point_sf = st_sf(geometry = st_as_sfc(edge_point))
  
  min(st_distance(edge_point_sf, city_centre))
}


distance_to_northern_edge = function(city_id, jetz_species_name) {
  distribution_northern_limit = distribution_northern_edges[distribution_northern_edges$SCI_NAME == jetz_species_name,]
  distance_to_edge(city_id, distribution_northern_limit)
}

distance_to_southern_edge = function(city_id, jetz_species_name) {
  distribution_southern_limit = distribution_southern_edges[distribution_southern_edges$SCI_NAME == jetz_species_name,]
  distance_to_edge(city_id, distribution_southern_limit)
}
```

```{r}
plot_distribution_limits(10, 'Columbina passerina')

distance_to_northern_edge(10, 'Columbina_passerina')
distance_to_southern_edge(10, 'Columbina_passerina')
```
```{r}
plot_distribution_limits(2503, 'Columba palumbus')

distance_to_northern_edge(2503, 'Columba_palumbus')
distance_to_southern_edge(2503, 'Columba_palumbus')
```
```{r}
plot_distribution_limits(1550, 'Columba palumbus')

distance_to_northern_edge(1550, 'Columba_palumbus')
distance_to_southern_edge(1550, 'Columba_palumbus')
```

```{r, warning = false}
write_csv(data.frame(city_id = NA, jetz_species_name = NA, distance_to_northern_edge = NA, distance_to_southern_edge = NA), filename(GEO_WORKING_OUTPUT_DIR, 'distance_to_distribution_edge.csv'))

for(index in 1:nrow(pools)) {
  row = pools[index,]
  
  distance_to_north = distance_to_northern_edge(row$city_id, row$jetz_species_name)
  distance_to_south = distance_to_southern_edge(row$city_id, row$jetz_species_name)
  
  write_csv(data.frame(city_id = row$city_id, jetz_species_name = row$jetz_species_name, distance_to_northern_edge = distance_to_north, distance_to_southern_edge = distance_to_south), filename(GEO_WORKING_OUTPUT_DIR, 'distance_to_distribution_edge.csv'), append = TRUE)
}
```
