==> scripts/data_prep/12_communities_for_analysis.Rmd <==
---
title: "Build Analysis Communities"
output: html_notebook
bibliography: ../ref.bib  
---

```{r}
source('../env.R')
```

# Build Communities
```{r}
raw_data = read_csv(filename(COMMUNITY_OUTPUT_DIR, 'recorded_species_with_spatial_data.csv'))
raw_data_trimmed_to_accepted_cities = raw_data[raw_data$buf_500m_percentage_total_city_area_surveyed > 10 & raw_data$total_city_locations > 10,]
length(unique(raw_data_trimmed_to_accepted_cities$city_id))
```

Trim out species in cities but not regional pool data
```{r}
pools = raw_data_trimmed_to_accepted_cities[!is.na(raw_data_trimmed_to_accepted_cities$seasonal),]
```

          
```{r}
write_csv(pools %>%
            dplyr::select(
              city_id,
              city_name,
              ebird_species_name,
              seasonal,
              presence,
              origin,
              buf_500m_percentage_survey_area_present
            ) %>% dplyr::rename(relative_abundance_proxy='buf_500m_percentage_survey_area_present'), filename(COMMUNITY_OUTPUT_DIR, 'communities_for_analysis.csv'))

city_data = pools %>%
            dplyr::select(
              city_id,
              city_name,
              total_city_checklists,
              total_city_locations,
              total_city_effort,
              total_city_area_m2,
              buf_500m_percentage_total_city_area_surveyed
            ) %>%
            distinct()

write_csv(city_data, filename(CITY_DATA_OUTPUT_DIR, 'city_effort.csv'))
```

Total checklists included in survey
```{r}
sum(city_data$total_city_checklists)
```

Total checklists which include at least one Columbidae:
```{r}
columbidae_checklists = read_csv(filename(EBIRD_WORKING_OUTPUT_DIR, 'urban_columbidae.csv'))
columbidae_checklists %>% filter(city_id %in% city_data$city_id) %>% select(checklist_id) %>% distinct() %>% count()
```

```{r}
city_data %>% arrange(total_city_area_m2) %>% mutate(total_city_area_km2 = total_city_area_m2 / (1000 * 1000)) %>% dplyr::select(city_name, total_city_area_km2)
```

Create shape file with selected cities
```{r}
initial_city_selection = st_read(filename(mkdir(GEO_WORKING_OUTPUT_DIR, 'cities'), 'initial_selection.shp'))
final_city_selection = initial_city_selection[initial_city_selection$city_id %in% city_data$city_id,] %>%
  left_join(city_data) %>% 
  dplyr::select(!c('total_city_area_m2')) %>%
  mutate(across(where(is.numeric), as.integer))
 
write_sf(final_city_selection, filename(CITY_DATA_OUTPUT_DIR, 'city_selection.shp'))
```

Plot our selected cities
```{r, fig.height=6, fig.width=6}
world_map = read_country_boundaries()

ggplot() +  
  geom_sf(data = world_map, aes(geometry = geometry)) +
  geom_sf(data = final_city_selection, aes(geometry = geometry, colour = buf_500m_percentage_total_city_area_surveyed)) + 
  theme(legend.position="bottom") + scale_colour_gradient2(
    low = "darkgreen",
    mid = "yellow",
    high = "red",
    midpoint = 3,
    space = "Lab",
    na.value = "grey50",
    guide = "colourbar",
    aesthetics = "colour",
  )
```


# Compare our communities to eBird Status and Trends abundance data

The data comes from the eBird status and trends data sets and is calculated from all the pixels that occur within a city polygon for each species.
```{r}
abundance_data = read_csv(filename(EBIRD_WORKING_OUTPUT_DIR, 'ebird_trends_abundance_test_data.csv'))
head(abundance_data)
```

```{r}
test_species = unique(abundance_data$species)
test_species
```


```{r, echo = FALSE, warning = FALSE}
abundance_data$max_mean_abundance = apply(abundance_data[,c('mean_breeding', 'mean_nonbreeding', 'mean_resident')], 1, max, na.rm = T)
abundance_data$max_mean_abundance[abundance_data$max_mean_abundance == -Inf] = NA
```

```{r}
analytical_pools = read_csv(filename(COMMUNITY_OUTPUT_DIR, 'communities_for_analysis.csv'))
comparison_data = left_join(analytical_pools, abundance_data[,c('city_id', 'species', 'max_mean_abundance')], by = c('ebird_species_name' = 'species', 'city_id' = 'city_id'))
comparison_data
```


```{r}
plot_data = comparison_data[!is.na(comparison_data$max_mean_abundance),]
summary(glm(formula = max_mean_abundance ~ poly(relative_abundance_proxy, 2), data = plot_data))
```

```{r}
pred_data <- predict(glm(formula = max_mean_abundance ~ poly(relative_abundance_proxy, 2), data = plot_data), se.fit=T)

number_of_degrees_out = 50
outside_se_data <- plot_data[plot_data$max_mean_abundance < pred_data$fit - number_of_degrees_out * pred_data$se.fit | plot_data$max_mean_abundance > pred_data$fit + number_of_degrees_out * pred_data$se.fit,]
outside_se_data$label = paste(outside_se_data$ebird_species_name, '\nin', outside_se_data$city_name)

comparison_data %>% 
  ggplot(aes(x = relative_abundance_proxy, y = max_mean_abundance)) +
    geom_point(colour = cyan) +
    xlab('Abundance Proxy\nCalculated from area of city covered by checklists') + 
    ylab('Mean relative abundance across cities\nfrom eBird Status & Trends') +
    labs(title = 'Abundance proxy vs eBird relative abundance estimate', subtitle='eBird Status & Trends abundance data') +
    geom_smooth(method = 'glm', formula = y ~ poly(x, 2), colour = magenta) + 
    geom_text_repel(aes(label = label), data = outside_se_data, size = 3, nudge_y = 1, color = red) +
    geom_point(data = outside_se_data, color = red) +
    theme_bw() +
    theme(legend.position = "none")

ggsave(filename(FIGURES_OUTPUT_DIR, 'abundance_proxy_to_ebird_trends_and_status_abundance_data.jpg'))
```
